requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
camelot-py==1.0.9
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.3.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pypdf==3.17.4
pypdfium2==4.30.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
starlette==0.47.2
tabulate==0.9.0
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websocket-client==1.8.0
wsproto==1.2.0
